separates Style from node module
