+++ podpac/__init__.py
@@ -14,7 +14,8 @@ version_info : TYPE
 # Public API
 from podpac.core.units import Units, UnitsDataArray, UnitsNode
 from podpac.core.coordinates import Coordinates, crange, clinspace
-from podpac.core.node import Node, Style
+from podpac.core.node import Node
+from podpac.core.style import Style
 from podpac.core.algorithm.algorithm import (
     Algorithm, Arithmetic, SinCoords)
 from podpac.core.algorithm.stats import (
